rewrite of nixos-config
Some of the main changes was added neovim plugins. However they are
still no fully tested and may not work as expected. Another big change
is the use of greetd as the displaymanager and/or loginmanager. I also
made Nvidia driver choose between the beta and the production driver
based on what is newer. The kernel had to be switched to 6.5 has 6.6 has
bug currently with Nvidia. Reading online it could be either Nvidia or
the Kernel that needs to be fixed. However, should try to find fix for
this or update the kernel in the future to latest or something else with
a release cycle. Enabled TPM in config however don't know if it works and may
disable in the future because of this. Added Theming folder which adds
fonts, cursors, icons, and GTK theming. However, still need to QT
theming and make sure theming is actually appiled. VSCode will work
well enought right now however I can not use unstable as it will crash.
The Stable VSCode will also crash but it is unlikely. There are more
changes which can be seen in the file diffrences.