Update to GPU config
Intel can now enable GuC and HuC for CPUs that do not enable them by default
The main change is now laptops work with nvidia prime offload. The idea will be that the program in the desktop manager that launches the various apps will be configured to have a lauching with Nvidia option when on laptop with a multi gpu setup.